+++ apps/user/src/apis/token/token.ts
@@ -1,4 +1,5 @@
-import TOKEN from '@/constants/token';
+import { TOKEN } from '@/constants/common/constant';
+
 import { Storage } from '../storage/storage';
 
 const authorization = () => {
