+++ apps/user/src/app/form/초안제출완료/초안제출완료.hooks.ts
@@ -1,10 +1,10 @@
-import ROUTES from '@/constants/routes';
-import { useFormStepState } from '@/hooks';
+import { ROUTES } from '@/constants/common/constant';
+import { useFormStepSetStore } from '@/store';
 import { useRouter } from 'next/navigation';
 
 export const useCTAButton = () => {
     const router = useRouter();
-    const { setFormStep } = useFormStepState();
+    const setFormStep = useFormStepSetStore();
 
     const handleGoMainPageButtonClick = () => {
         router.push(ROUTES.MAIN);
