+++ apps/user/src/components/common/Footer/Footer.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/navigation';
 import { Column, Row, Link } from '@maru/ui';
 import { font, color } from '@maru/theme';
 import { flex } from '@maru/utils';
-import ROUTES from '@/constants/routes';
+import { ROUTES } from '@/constants/common/constant';
 import Image from 'next/image';
 import styled from 'styled-components';
