+++ apps/user/src/components/main/Dday/Dday.hooks.ts
@@ -6,7 +6,7 @@ import {
     최종_합격_발표,
     제출_마감_날짜,
     제출_시작_날짜,
-} from '@/constants/service/submitTime';
+} from '@/constants/common/constant';
 import type { ButtonOptionType } from '@maru/ui';
 import { formatDay } from '@/utils';
 import isBetween from 'dayjs/plugin/isBetween';
