+++ apps/user/src/components/main/Faq/MainFaq.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/navigation';
-import ROUTES from '@/constants/routes';
+import { ROUTES } from '@/constants/common/constant';
 import { Link } from '@maru/ui';
 import { color, font } from '@maru/theme';
 import { flex } from '@maru/utils';
