+++ apps/user/src/constants/form/constant.ts
@@ -1,0 +1,14 @@
+export const REGULAR_TYPE_DEFAULT_SCORE = 80;
+export const SPECIAL_TYPE_DEFAULT_SCORE = 48;
+
+export const DEFAULT_ATTENDANCE_SCORE = 14;
+export const MAX_ABSENCE_COUNT = 18;
+export const MIN_ATTENDANCE_SCORE = 0;
+export const MAX_ATTENDANCE_SCORE = 18;
+
+export const DEFAULT_VOLUNTEER_SCORE = 14;
+
+export const MIN_VOLUNTEER_TIME = 7;
+export const MIN_VOLUNTEER_SCORE = 0;
+export const MAX_VOLUNTEER_TIME = 15;
+export const MAX_VOLUNTEER_SCORE = 18;
