--- apps/user/src/constants/key.ts
@@ -1,12 +1,0 @@
-const KEY = {
-    NOTICE_LIST: 'useNoticeList',
-    NOTICE_DETAIL: 'useNoticeDetail',
-    FAQ_CATEGORY_LIST: 'useFaqCategoryListQuery',
-    FAQ_LIST: 'useFaqListQuery',
-    FORM_SCHOOL_LIST: 'useFormSchoolListQuery',
-    EXPORT_FORM: 'useExportForm',
-    SAVE_FORM: 'useSaveForm',
-    USER: 'useUser',
-} as const;
-
-export default KEY;
