--- apps/user/src/constants/routes.ts
@@ -1,10 +1,0 @@
-const ROUTES = {
-    MAIN: '/',
-    FAQ: '/faq',
-    FORM: '/form',
-    LOGIN: '/login',
-    NOTICE: '/notice',
-    SIGNUP: '/signup',
-} as const;
-
-export default ROUTES;
