--- apps/user/src/constants/token.ts
@@ -1,6 +1,0 @@
-const TOKEN = {
-    ACCESS: 'access-token',
-    REFRESH: 'refresh-token',
-} as const;
-
-export default TOKEN;
