+++ apps/user/src/services/user/queries.ts
@@ -1,6 +1,5 @@
 import { Storage } from '@/apis/storage/storage';
-import KEY from '@/constants/key';
-import TOKEN from '@/constants/token';
+import { TOKEN, KEY } from '@/constants/common/constant';
 import { useQuery } from '@tanstack/react-query';
 import { getUser } from './api';
