+++ apps/user/src/store/form/subject.ts
@@ -1,10 +1,10 @@
-import { SUBJECT_LIST_DATA } from '@/constants/service/form';
+import { SUBJECT_LIST } from '@/constants/form/initial';
 import { Subject } from '@/types/form/client';
-import { atom, useRecoilState, useRecoilValue } from 'recoil';
+import { atom, useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 
 const subjectListAtomState = atom<Subject[]>({
     key: 'subject-list',
-    default: [...SUBJECT_LIST_DATA],
+    default: [...SUBJECT_LIST],
 });
 
 const newSubjectListAtomState = atom<Subject[]>({
@@ -12,10 +12,10 @@ const newSubjectListAtomState = atom<Subject[]>({
     default: [],
 });
 
-export const useNewSubjectStore = () => useRecoilState(newSubjectListAtomState);
-export const useNewSubjectValueStore = () => useRecoilValue(newSubjectListAtomState);
-export const useNewSubjectSetStore = () => useRecoilValue(newSubjectListAtomState);
-
 export const useSubjectStore = () => useRecoilState(subjectListAtomState);
 export const useSubjectValueStore = () => useRecoilValue(subjectListAtomState);
-export const useSubjectSetStore = () => useRecoilValue(subjectListAtomState);
+export const useSubjectSetStore = () => useSetRecoilState(subjectListAtomState);
+
+export const useNewSubjectStore = () => useRecoilState(newSubjectListAtomState);
+export const useNewSubjectValueStore = () => useRecoilValue(newSubjectListAtomState);
+export const useNewSubjectSetStore = () => useSetRecoilState(newSubjectListAtomState);
